+++ Semana_4/node/src/modules/users/services/SendForgotPasswordEmailService.ts
@@ -1,7 +1,7 @@
 import IMailProvider from "@shared/container/providers/SendEmailProvider/models/IMailProvider";
 import AppError from "@shared/errors/AppError";
 import IUserRepository from "@users/repositories/IUsersRepository";
-import IUserTokenRepository from '@users/repositories/IUserTokenRepository'
+import IUserTokenRepository from "@users/repositories/IUserTokenRepository";
 import { inject, injectable } from "tsyringe";
 
 interface IRequestDTO {
@@ -26,10 +26,24 @@ class SendForgotPasswordEmailService {
     if (!user)
       throw new AppError("User not found - The email must be valid!", 404);
 
-    let {token} = await this.userTokenRepository.generate(user.id)
+    let { token } = await this.userTokenRepository.generate(user.id);
 
     let body = `Pedido de recuperação de senha recebido com sucesso!\nUtilize o token: ${token} para recuperação!`;
-    await this.mailProvider.sendMail(email, body);
+    await this.mailProvider.sendMail({
+      subject: `Recuperação de Senha!`,
+      to: {
+        email: user.email,
+        name: user.name,
+      },
+      from: { email: "equipe@gobarber.com", name: "Equipe GoBarber" },
+      templateData: {
+        template: "Olá, {{name}}: {{token}}",
+        variables: {
+          name: user.name,
+          token,
+        },
+      },
+    });
   }
 }