+++ Semana_4/node/src/shared/container/providers/SendEmailProvider/implementations/EtherealMailProvider.ts
@@ -1,10 +1,16 @@
 import IMailProvider from "@shared/container/providers/SendEmailProvider/models/IMailProvider";
-
+import IMailTemplateProvider from "@shared/container/providers/MailTemplateProvider/models/IMailTemplateProvider";
+import ISendEmailDTO from "../dtos/ISendEmailDTO";
 import nodemailer, { Transporter } from "nodemailer";
+import { inject, injectable } from "tsyringe";
 
+@injectable()
 export default class EtherealMailProvider implements IMailProvider {
   private client: Transporter;
-  constructor() {
+  constructor(
+    @inject("MailTemplateProvider")
+    private mailTemplateProvider: IMailTemplateProvider
+  ) {
     nodemailer.createTestAccount().then((account) => {
       let transporter = nodemailer.createTransport({
         host: account.smtp.host,
@@ -18,12 +24,20 @@ export default class EtherealMailProvider implements IMailProvider {
       this.client = transporter;
     });
   }
-  public async sendMail(to: string, body: string): Promise<void> {
+  public async sendMail({
+    to,
+    subject,
+    templateData,
+    from,
+  }: ISendEmailDTO): Promise<void> {
     let message = await this.client.sendMail({
-      to,
-      from: "Equipe GoBarber <equipe@gobarber.com.br>",
-      subject: "Recuperação de Senha!",
-      text: body,
+      to: `${to.name} <${to.email}>`,
+      from: {
+        name: from?.name || "Equipe GoBarber",
+        address: from?.email || "equipe@gobarber.com.br",
+      },
+      subject,
+      html: await this.mailTemplateProvider.parse(templateData),
     });
 
     console.log("Message sent: %s", message.messageId);