+++ Semana_4/node/src/shared/container/providers/index.ts
@@ -7,13 +7,21 @@ import IMailProvider from "@shared/container/providers/SendEmailProvider/models/
 import EtherealMailProvider from "@shared/container/providers/SendEmailProvider/implementations/EtherealMailProvider";
 //import IStorageProvider from '@shared/container/providers/SendEmailProvider/implementations/'
 
-container.registerInstance<IMailProvider>(
-  "MailProvider",
-  new EtherealMailProvider()
-);
+import IMailTemplateProvider from "@shared/container/providers/MailTemplateProvider/models/IMailTemplateProvider";
+import HandleBarsMailTemplateProvider from "@shared/container/providers/MailTemplateProvider/implementations/HandleBarsMailTemplateProvider";
 
 container.registerSingleton<IStorageProvider>(
   "StorageProvider",
   DiskStorageProvider
+);
+
+container.registerSingleton<IMailTemplateProvider>(
+  "MailTemplateProvider",
+  HandleBarsMailTemplateProvider
+);
+
+container.registerInstance<IMailProvider>(
+  "MailProvider",
+  container.resolve(EtherealMailProvider)
 );
 //container.registerSingleton<IMailProvider>('MailProvider',MailProvider)